Fix notes when playing chords
Chords were playing the wrong notes because the note name didn't
include accidentals. Because there are also problems detecting,
say, Eb as equivalent to D#, change the way we refer to notes when
playing them, to use the .chroma() value, which is the position of
the note in the scale, and thus the same for the same note,
regardless of nomenclature ("Eb" vs. "D#").