Oh, boy. Finally fix the silent note bug.
It seems that the problem was actually one of the divs showing the
note and/or the note role. They were getting the click event, and
thus the key wasn't getting it, and the note wasn't being played.
Solve the problem with the good old pointer-events CSS property.

I hope I'm right and this really is the end of it.