Add an animation when playing from the keyboard
When playing the keys with the mouse, there's the feedback from
pressing the "button". However, there was no feedback when playing
from the keyboard, which is more important because it's not
evident which note is being played. Hence, add a 300ms animation
when playing from the keyboard, so it's clear what is happening.